+++ extensions/raycast-wallpaper/src/components/raycast-wallpaper-grid.tsx
@@ -1,10 +1,10 @@
-import { getPreferenceValues, Grid, Icon } from "@raycast/api";
+import { Color, Grid, Icon } from "@raycast/api";
 import React from "react";
 import { RaycastWallpaperWithInfo } from "../types/types";
 import { RaycastWallpaperEmptyView } from "./raycast-wallpaper-empty-view";
-import { Preferences } from "../types/preferences";
 import { ActionOnRaycastWallpaper } from "./action-on-raycast-wallpaper";
-import { getThumbnailUrl } from "../utils/common-utils";
+import { capitalizeFirstLetter, getThumbnailUrl } from "../utils/common-utils";
+import { columns, layout, respectAppearance } from "../types/preferences";
 
 export function RaycastWallpaperGrid(props: {
   raycastWallpapers: RaycastWallpaperWithInfo[];
@@ -12,13 +12,12 @@ export function RaycastWallpaperGrid(props: {
   selectedItem: string;
   setSelectedItem: React.Dispatch<React.SetStateAction<string>>;
 }) {
-  const preferences = getPreferenceValues<Preferences>();
   const { raycastWallpapers, setRefresh, selectedItem, setSelectedItem } = props;
 
   return (
     <Grid
       isLoading={raycastWallpapers.length === 0}
-      columns={parseInt(preferences.columns)}
+      columns={parseInt(columns)}
       aspectRatio={"16/9"}
       fit={Grid.Fit.Fill}
       selectedItemId={selectedItem}
@@ -29,8 +28,15 @@ export function RaycastWallpaperGrid(props: {
       }}
       searchBarPlaceholder={"Search wallpapers..."}
     >
-      <RaycastWallpaperEmptyView layout={preferences.layout} />
+      <RaycastWallpaperEmptyView layout={layout} />
       {raycastWallpapers.map((value, index) => {
+        const appearanceIcon = value.appearance == "light" ? Icon.Sun : Icon.Moon;
+        const accessories = respectAppearance
+          ? {
+              icon: { source: appearanceIcon, tintColor: Color.SecondaryText },
+              tooltip: capitalizeFirstLetter(value.appearance),
+            }
+          : undefined;
         return (
           <Grid.Item
             id={index + ""}
@@ -45,7 +51,14 @@ export function RaycastWallpaperGrid(props: {
                 setSelectedItem={setSelectedItem}
               />
             }
-            accessory={value.exclude ? { icon: Icon.XMarkTopRightSquare, tooltip: "Excluded From Auto Switch" } : {}}
+            accessory={
+              value.exclude
+                ? {
+                    icon: { source: Icon.XMarkTopRightSquare, tintColor: Color.SecondaryText },
+                    tooltip: "Excluded From Auto Switch",
+                  }
+                : accessories
+            }
           />
         );
       })}