+++ extensions/raycast-wallpaper/src/components/raycast-wallpaper-list.tsx
@@ -1,10 +1,10 @@
-import { getPreferenceValues, Icon, List } from "@raycast/api";
+import { Icon, List } from "@raycast/api";
 import React from "react";
 import { RaycastWallpaperWithInfo } from "../types/types";
 import { RaycastWallpaperEmptyView } from "./raycast-wallpaper-empty-view";
-import { Preferences } from "../types/preferences";
 import { ActionOnRaycastWallpaper } from "./action-on-raycast-wallpaper";
-import { getThumbnailUrl } from "../utils/common-utils";
+import { capitalizeFirstLetter, getThumbnailUrl } from "../utils/common-utils";
+import { layout, respectAppearance } from "../types/preferences";
 
 export function RaycastWallpaperList(props: {
   raycastWallpapers: RaycastWallpaperWithInfo[];
@@ -12,7 +12,6 @@ export function RaycastWallpaperList(props: {
   selectedItem: string;
   setSelectedItem: React.Dispatch<React.SetStateAction<string>>;
 }) {
-  const preferences = getPreferenceValues<Preferences>();
   const { raycastWallpapers, setRefresh, selectedItem, setSelectedItem } = props;
 
   return (
@@ -27,8 +26,12 @@ export function RaycastWallpaperList(props: {
       }}
       searchBarPlaceholder={"Search wallpapers"}
     >
-      <RaycastWallpaperEmptyView layout={preferences.layout} />
+      <RaycastWallpaperEmptyView layout={layout} />
       {raycastWallpapers.map((value, index) => {
+        const appearanceIcon = value.appearance === "light" ? Icon.Sun : Icon.Moon;
+        const accessories = respectAppearance
+          ? [{ icon: appearanceIcon, tooltip: capitalizeFirstLetter(value.appearance) }]
+          : [];
         return (
           <List.Item
             id={index + ""}
@@ -36,7 +39,7 @@ export function RaycastWallpaperList(props: {
             icon={{ source: getThumbnailUrl(value.url) }}
             title={value.title}
             accessories={
-              value.exclude ? [{ icon: Icon.XMarkTopRightSquare, tooltip: "Excluded From Auto Switch" }] : []
+              value.exclude ? [{ icon: Icon.XMarkTopRightSquare, tooltip: "Excluded From Auto Switch" }] : accessories
             }
             detail={<List.Item.Detail isLoading={false} markdown={`![](${getThumbnailUrl(value.url)})`} />}
             actions={